+++ packages/jala/lib/src/file_export_sink.dart
@@ -1,0 +1,2 @@
+export 'file_export_sink_stub.dart'
+    if (dart.library.io) 'file_export_sink_io.dart';